refactor(lsposed): extract sanitizers + wire up ConnectivityService hooks
Split the VPN-stripping logic out of HookEntry's writeToParcel handlers
into a new NetworkSanitizers object with three pure functions
(sanitizeNetworkCapabilities, sanitizeNetworkInfo, sanitizeLinkProperties).
Each returns a modified copy or null when no change is needed — no
in-place mutation of objects the ConnectivityService may be sharing
across threads.

Wire the previously dormant ConnectivityServiceHooks.install() from
HookEntry now that the sanitizers are callable from both code paths.
The writeToParcel hooks stay in place as a defensive fallback for
callback/broadcast paths the server-side hooks may not yet cover.

Array results (NetworkInfo[], Network[], NetworkCapabilities[]) and
Network handle hiding are still TODO in ConnectivityServiceHooks.
Basic coverage today: NetworkInfo getters, NetworkCapabilities,
LinkProperties scalar getters.
