+++ lsposed/app/src/main/kotlin/dev/okhsunrog/vpnhide/HookEntry.kt
@@ -3,7 +3,6 @@ package dev.okhsunrog.vpnhide
 import android.net.LinkProperties
 import android.net.NetworkCapabilities
 import android.net.NetworkInfo
-import android.net.RouteInfo
 import android.os.Binder
 import de.robv.android.xposed.IXposedHookLoadPackage
 import de.robv.android.xposed.XC_MethodHook
@@ -50,7 +49,7 @@ class HookEntry : IXposedHookLoadPackage {
 
         if (hookInstalled.compareAndSet(false, true)) {
             XposedBridge.log("VpnHide: system_server detected, installing Binder hooks")
-            installSystemServerHooks()
+            installSystemServerHooks(lpparam.classLoader)
             tryHook("PackageVisibility") { PackageVisibilityHooks.install(lpparam.classLoader) }
             writeHookStatusFile()
         }
@@ -68,90 +67,10 @@ class HookEntry : IXposedHookLoadPackage {
     }
 
     // ------------------------------------------------------------------
-    //  Helpers
+    //  Helpers — sanitise logic lives in NetworkSanitizers so both the
+    //  legacy writeToParcel hooks below and the new ConnectivityService
+    //  getter hooks share one implementation.
     // ------------------------------------------------------------------
-
-    private fun isVpnInterfaceName(name: String): Boolean {
-        if (name.isEmpty()) return false
-        val n = name.lowercase()
-        return n.startsWith("tun") ||
-            n.startsWith("ppp") ||
-            n.startsWith("tap") ||
-            n.startsWith("wg") ||
-            n.startsWith("ipsec") ||
-            n.startsWith("xfrm") ||
-            n.startsWith("utun") ||
-            n.startsWith("l2tp") ||
-            n.startsWith("gre") ||
-            n.contains("vpn")
-    }
-
-    private fun sanitizeLinkProperties(copy: LinkProperties): Boolean {
-        var modified = false
-
-        val ifaceName = XposedHelpers.getObjectField(copy, "mIfaceName") as? String
-        if (ifaceName != null && isVpnInterfaceName(ifaceName)) {
-            XposedHelpers.setObjectField(copy, "mIfaceName", null)
-            modified = true
-        }
-
-        try {
-            @Suppress("UNCHECKED_CAST")
-            val routesField = XposedHelpers.getObjectField(copy, "mRoutes") as? MutableList<RouteInfo>
-            if (routesField != null) {
-                val filtered =
-                    routesField.filterNot { route ->
-                        val routeIface = route.`interface`
-                        routeIface != null && isVpnInterfaceName(routeIface)
-                    }
-                if (filtered.size != routesField.size) {
-                    routesField.clear()
-                    routesField.addAll(filtered)
-                    modified = true
-                }
-            }
-        } catch (t: Throwable) {
-            XposedBridge.log("VpnHide: failed to sanitize mRoutes: ${t.message}")
-        }
-
-        try {
-            @Suppress("UNCHECKED_CAST")
-            val stacked = XposedHelpers.getObjectField(copy, "mStackedLinks") as? MutableMap<String, LinkProperties>
-            if (stacked != null && stacked.isNotEmpty()) {
-                val filtered = LinkedHashMap<String, LinkProperties>()
-                for ((key, value) in stacked) {
-                    val stackedCopy =
-                        try {
-                            val ctor = LinkProperties::class.java.getDeclaredConstructor(LinkProperties::class.java)
-                            ctor.isAccessible = true
-                            ctor.newInstance(value) as LinkProperties
-                        } catch (_: Throwable) {
-                            value
-                        }
-                    val stackedModified = sanitizeLinkProperties(stackedCopy)
-                    val stackedIface = XposedHelpers.getObjectField(stackedCopy, "mIfaceName") as? String
-                    if (stackedIface == null && stackedCopy.routes.isEmpty()) {
-                        if (stackedModified || isVpnInterfaceName(key)) {
-                            modified = true
-                        } else {
-                            filtered[key] = stackedCopy
-                        }
-                    } else {
-                        if (stackedModified || stackedCopy !== value) modified = true
-                        filtered[key] = stackedCopy
-                    }
-                }
-                if (filtered.size != stacked.size || modified) {
-                    stacked.clear()
-                    stacked.putAll(filtered)
-                }
-            }
-        } catch (t: Throwable) {
-            XposedBridge.log("VpnHide: failed to sanitize mStackedLinks: ${t.message}")
-        }
-
-        return modified
-    }
 
     // ==================================================================
     //  system_server hooks — per-UID Binder filtering
@@ -205,10 +124,19 @@ class HookEntry : IXposedHookLoadPackage {
         systemServerTargetUids = null
     }
 
-    private fun installSystemServerHooks() {
+    private fun installSystemServerHooks(classLoader: ClassLoader) {
         tryHook("NC.writeToParcel") { hookNCWriteToParcel() }
         tryHook("NI.writeToParcel") { hookNIWriteToParcel() }
         tryHook("LP.writeToParcel") { hookLPWriteToParcel() }
+        tryHook("ConnectivityService getters") {
+            ConnectivityServiceHooks.install(
+                classLoader = classLoader,
+                isTargetCaller = ::isTargetCaller,
+                sanitizeNetworkCapabilities = NetworkSanitizers::sanitizeNetworkCapabilities,
+                sanitizeNetworkInfo = NetworkSanitizers::sanitizeNetworkInfo,
+                sanitizeLinkProperties = NetworkSanitizers::sanitizeLinkProperties,
+            )
+        }
         tryHook("FileObserver") { watchTargetUidsFile() }
     }
 
@@ -260,11 +188,13 @@ class HookEntry : IXposedHookLoadPackage {
     }
 
     /**
-     * Hook NetworkCapabilities.writeToParcel in system_server.
-     * For target UIDs, creates a copy with VPN stripped and writes
-     * the copy to the Parcel instead of the original. The original
-     * object is never mutated, avoiding race conditions with
-     * ConnectivityService threads.
+     * Hook NetworkCapabilities.writeToParcel — defensive fallback for
+     * callback/broadcast paths the ConnectivityService getter hooks may
+     * not cover. For target UIDs, writes a VPN-stripped copy to the
+     * Parcel instead of the original; the original object is never
+     * mutated.
+     *
+     * Slated for removal once server-side hooks prove sufficient.
      */
     private fun hookNCWriteToParcel() {
         val writingCopy = ThreadLocal<Boolean>()
@@ -276,31 +206,16 @@ class HookEntry : IXposedHookLoadPackage {
             object : XC_MethodHook() {
                 override fun beforeHookedMethod(param: MethodHookParam) {
                     if (writingCopy.get() == true) return
-                    if (!loadTargetUids().contains(Binder.getCallingUid())) return
+                    if (!isTargetCaller()) return
 
                     val nc = param.thisObject as NetworkCapabilities
                     try {
-                        val transportTypes = XposedHelpers.getLongField(nc, "mTransportTypes")
-                        val vpnBit = 1L shl TRANSPORT_VPN
-                        if (transportTypes and vpnBit == 0L) return
-
-                        val copy = NetworkCapabilities(nc)
-                        XposedHelpers.setLongField(copy, "mTransportTypes", transportTypes and vpnBit.inv())
-                        val caps = XposedHelpers.getLongField(copy, "mNetworkCapabilities")
-                        XposedHelpers.setLongField(copy, "mNetworkCapabilities", caps or (1L shl NET_CAPABILITY_NOT_VPN))
-                        try {
-                            val ti = XposedHelpers.getObjectField(copy, "mTransportInfo")
-                            if (ti != null && ti.javaClass.name == "android.net.VpnTransportInfo") {
-                                XposedHelpers.setObjectField(copy, "mTransportInfo", null)
-                            }
-                        } catch (_: Throwable) {
-                        }
-
+                        val sanitized = NetworkSanitizers.sanitizeNetworkCapabilities(nc) ?: return
                         val parcel = param.args[0] as android.os.Parcel
                         val flags = param.args[1] as Int
                         writingCopy.set(true)
                         try {
-                            copy.writeToParcel(parcel, flags)
+                            sanitized.writeToParcel(parcel, flags)
                         } finally {
                             writingCopy.set(false)
                         }
@@ -314,10 +229,6 @@ class HookEntry : IXposedHookLoadPackage {
         XposedBridge.log("VpnHide: hooked NetworkCapabilities.writeToParcel")
     }
 
-    /**
-     * Hook NetworkInfo.writeToParcel — disguise VPN NetworkInfo for target callers.
-     * Creates a copy with type changed from VPN to WIFI, writes the copy.
-     */
     @Suppress("DEPRECATION")
     private fun hookNIWriteToParcel() {
         val writingCopy = ThreadLocal<Boolean>()
@@ -332,27 +243,12 @@ class HookEntry : IXposedHookLoadPackage {
                     if (!isTargetCaller()) return
                     val ni = param.thisObject as NetworkInfo
                     try {
-                        val type = XposedHelpers.getIntField(ni, "mNetworkType")
-                        if (type != TYPE_VPN) return
-
-                        val ctor =
-                            NetworkInfo::class.java.getDeclaredConstructor(
-                                Integer.TYPE,
-                                Integer.TYPE,
-                                String::class.java,
-                                String::class.java,
-                            )
-                        ctor.isAccessible = true
-                        val copy = ctor.newInstance(TYPE_WIFI, 0, "WIFI", "") as NetworkInfo
-                        XposedHelpers.setIntField(copy, "mState", XposedHelpers.getIntField(ni, "mState"))
-                        XposedHelpers.setIntField(copy, "mDetailedState", XposedHelpers.getIntField(ni, "mDetailedState"))
-                        XposedHelpers.setBooleanField(copy, "mIsAvailable", XposedHelpers.getBooleanField(ni, "mIsAvailable"))
-
+                        val sanitized = NetworkSanitizers.sanitizeNetworkInfo(ni) ?: return
                         val parcel = param.args[0] as android.os.Parcel
                         val flags = param.args[1] as Int
                         writingCopy.set(true)
                         try {
-                            copy.writeToParcel(parcel, flags)
+                            sanitized.writeToParcel(parcel, flags)
                         } finally {
                             writingCopy.set(false)
                         }
@@ -366,11 +262,6 @@ class HookEntry : IXposedHookLoadPackage {
         XposedBridge.log("VpnHide: hooked NetworkInfo.writeToParcel")
     }
 
-    /**
-     * Hook LinkProperties.writeToParcel — clear VPN interface name and
-     * routes for target callers. Creates a copy to avoid mutating the
-     * original object shared by ConnectivityService threads.
-     */
     private fun hookLPWriteToParcel() {
         val writingCopy = ThreadLocal<Boolean>()
         XposedHelpers.findAndHookMethod(
@@ -384,16 +275,12 @@ class HookEntry : IXposedHookLoadPackage {
                     if (!isTargetCaller()) return
                     val lp = param.thisObject as LinkProperties
                     try {
-                        val ctor = LinkProperties::class.java.getDeclaredConstructor(LinkProperties::class.java)
-                        ctor.isAccessible = true
-                        val copy = ctor.newInstance(lp) as LinkProperties
-                        if (!sanitizeLinkProperties(copy)) return
-
+                        val sanitized = NetworkSanitizers.sanitizeLinkProperties(lp) ?: return
                         val parcel = param.args[0] as android.os.Parcel
                         val flags = param.args[1] as Int
                         writingCopy.set(true)
                         try {
-                            copy.writeToParcel(parcel, flags)
+                            sanitized.writeToParcel(parcel, flags)
                         } finally {
                             writingCopy.set(false)
                         }
@@ -408,10 +295,6 @@ class HookEntry : IXposedHookLoadPackage {
     }
 
     companion object {
-        private const val TRANSPORT_VPN = 4
-        private const val NET_CAPABILITY_NOT_VPN = 15
-        private const val TYPE_VPN = 17
-        private const val TYPE_WIFI = 1
         const val HOOK_STATUS_FILE = "/data/system/vpnhide_hook_active"
     }
 }
